Move listing into index page component
